+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # CHANGELOG
 
+## v0.1.3
+
+Changed:
+  - Improved build process.
+
+## v0.1.2
+
+Fixed:
+    - `JSString.__str__` correctly returns string value.
+    - Export `JSValue` type.
+
 ## v0.1.1
 
 Changed:
     - Renamed `JSContext.load_script` to `JSContext.load`.
-
 
 ## v0.1.0